Aggiunti test per le classi OrdineDAO, RecensioneDAO, ServizioDAO. Modificato tipo del voto, int-->float. Il metodo setStatoOrdine era errato, sostituito statoOrdine con nuovostatoOrdine dove necessario. Cambiato l'ordine di due set in Update in OrdineDAO. Aggiunto controllo per verificare l'appartenenza al giusto range riguardo il voto nelle recensioni. Aggiunto controllo sul prezzo, condzione: prezzo>0.

+++ src/Controller/ControllerUtente.java
@@ -163,7 +163,7 @@ public class ControllerUtente {
     }
 
     // Metodi per la gestione delle recensioni
-    public boolean effettuaRecensione(int idVenditore, int voto, String testo) {
+    public boolean effettuaRecensione(int idVenditore, float voto, String testo) {
         // Controllo che il voto sia compreso tra 1 e 5
         if (voto < 1 || voto > 5)
             return false;
